+++ library/Configuration/Exception.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * This file is part of slick/configuration package
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\Configuration;
+
+/**
+ * General configuration Exception
+ *
+ * @package Slick\Configuration
+ * @author  Filipe Silva <silvam.filipe@gmail.com>
+ */
+interface Exception
+{
+
+}