+++ library/Configuration/Exception/InvalidArgumentException.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * This file is part of slick/configuration package
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\Configuration\Exception;
+
+use LogicException;
+use Slick\Configuration\Exception;
+
+/**
+ * Invalid Argument Exception
+ *
+ * @package Slick\Configuration\Exception
+ */
+class InvalidArgumentException extends LogicException implements Exception
+{
+
+}